+++ src/Sulu/Bundle/WebsiteBundle/Twig/Meta/MetaTwigExtension.php
@@ -145,7 +145,11 @@ class MetaTwigExtension extends \Twig_Extension
     {
         $url = $this->contentPath->getContentPath($url, $webspaceKey, $locale);
 
-        return sprintf('<link rel="alternate" href="%s" hreflang="%s" />', $url, !$default ? $locale : 'x-default');
+        return sprintf(
+            '<link rel="alternate" href="%s" hreflang="%s" />',
+            $url,
+            !$default ? str_replace('_', '-', $locale) : 'x-default'
+        );
     }
 
     /**